git-svn-id: https://repo.acin.tuwien.ac.at/v4r/stable/strandsv4r@8398 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e

+++ faat_framework/3d_rec_framework/include/faat_pcl/3d_rec_framework/pipeline/impl/global_nn_recognizer_cvfh.hpp
@@ -406,7 +406,7 @@ template<template<class > class Distance, typename PointInT, typename FeatureT>
 
             //gather NN-search results
             double score = 0;
-            for (int i = 0; i < NN_; ++i)
+            for (size_t i = 0; i < NN_; ++i)
             {
               score = distances[0][i];
               index_score is;
